refactor: Merge `clap_generate_fig` into `clap_generate`
