+++ clap_generate/src/generators/shells/mod.rs
@@ -1,11 +1,13 @@
 mod bash;
 mod elvish;
+mod fig;
 mod fish;
 mod powershell;
 mod zsh;
 
 pub use bash::Bash;
 pub use elvish::Elvish;
+pub use fig::Fig;
 pub use fish::Fish;
 pub use powershell::PowerShell;
 pub use zsh::Zsh;
