+++ clap_generate/tests/generate_completions.rs
@@ -15,6 +15,7 @@ fn generate_completions() {
 
     generate(Bash, &mut app, "test_app", &mut io::sink());
     generate(Fish, &mut app, "test_app", &mut io::sink());
+    generate(Fig, &mut app, "test_app", &mut io::sink());
     generate(PowerShell, &mut app, "test_app", &mut io::sink());
     generate(Elvish, &mut app, "test_app", &mut io::sink());
     generate(Zsh, &mut app, "test_app", &mut io::sink());
