--- clap_generate_fig/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "clap_generate_fig"
-version = "3.0.0-rc.4"
-edition = "2018"
-include = [
-	"src/**/*",
-	"Cargo.toml",
-	"LICENSE-*",
-	"README.md"
-]
-description = "A generator library used with clap for Fig completion scripts"
-repository = "https://github.com/clap-rs/clap/tree/master/clap_generate_fig"
-documentation = "https://docs.rs/clap_generate_fig"
-keywords = [
-	"clap",
-	"cli",
-	"generate",
-	"completion",
-	"fig",
-]
-categories = ["command-line-interface"]
-license = "MIT OR Apache-2.0"
-readme = "README.md"
-
-[lib]
-bench = false
-
-[dependencies]
-clap = { path = "../", version = "=3.0.0-rc.4", default-features = false, features = ["std"] }
-clap_generate = { path = "../clap_generate", version = "=3.0.0-rc.4"}
-
-[dev-dependencies]
-pretty_assertions = "1.0"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
