--- clap_generate_fig/README.md
@@ -1,9 +1,0 @@
-# clap_generate_fig
-
-Generates [Fig](https://github.com/withfig/autocomplete) completions for [`clap`](https://github.com/clap-rs/clap) based CLIs
-
-<!-- * [Documentation][docs] -->
-- [Questions & Discussions](https://github.com/clap-rs/clap/discussions)
-- [CONTRIBUTING](../CONTRIBUTING.md)
-
-<!-- [docs]: https://docs.rs/clap_generate_fig -->
